chore (views) split styles from views
