+++ src/views/MainView/styles.js
@@ -1,0 +1,26 @@
+import { StyleSheet }       from 'react-native';
+import Global               from '../../util/globals';
+
+const MARKER_IMAGE_SIZE = 32;
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    flexDirection: 'column',
+    backgroundColor: Global.COLOR.LIGHT_GREY,
+  },
+  mapContainer: {
+    flex: 1,
+    flexDirection: 'column',
+    alignItems: 'stretch',
+  },
+  map: {
+    flex: 1,
+  },
+  markerImage: {
+    height: MARKER_IMAGE_SIZE,
+    width: MARKER_IMAGE_SIZE,
+  }
+});
+
+export default styles
