+++ src/views/SendServiceRequestView/index.js
@@ -2,13 +2,11 @@ import React, { Component, } from 'react';
 import rebound from 'rebound';
 import {
   View,
-  StyleSheet,
   Image,
   Text,
   TextInput,
   TouchableWithoutFeedback,
   BackAndroid,
-  Dimensions,
   Platform,
   Animated,
   NativeModules,
@@ -24,33 +22,32 @@ import ImageResizer                 from 'react-native-image-resizer';
 import Toast                        from 'react-native-simple-toast';
 import KeyboardSpacer               from 'react-native-keyboard-spacer';
 import Spinner                      from 'react-native-loading-spinner-overlay';
-import FloatingActionButton         from '../components/FloatingActionButton';
-import NativePicker                 from '../components/NativePicker';
-import Navbar                       from '../components/Navbar';
-import Menu                         from '../components/Menu';
-import Thumbnail                    from '../components/Thumbnail';
-import showAlert                    from '../components/Alert';
-import SendServiceRequestMap        from '../components/SendServiceRequestMap';
-import SendServiceRequestForm       from '../components/SendServiceRequestForm';
-import SendServiceRequestAttachment from '../components/SendServiceRequestAttachment';
-import makeRequest                  from '../util/requests';
-import serviceRequestModels         from '../util/models';
-import Global                       from '../util/globals';
-import Util                         from '../util/util';
-import Config                       from '../config';
-import transSendServiceRequest      from '../translations/sendServiceRequest';
-import transError                   from '../translations/errors';
-import checkboxIcon                 from '../img/check.png';
-import sendEnabledIcon              from '../img/send_enabled.png';
-import sendDisabledIcon             from '../img/send_disabled.png';
-import markerIcon                   from '../img/location_marker.png';
-import checkIcon                    from '../img/check.png';
-import backIcon                     from '../img/back.png';
+import FloatingActionButton         from '../../components/FloatingActionButton';
+import NativePicker                 from '../../components/NativePicker';
+import Navbar                       from '../../components/Navbar';
+import Menu                         from '../../components/Menu';
+import Thumbnail                    from '../../components/Thumbnail';
+import showAlert                    from '../../components/Alert';
+import SendServiceRequestMap        from '../../components/SendServiceRequestMap';
+import SendServiceRequestForm       from '../../components/SendServiceRequestForm';
+import SendServiceRequestAttachment from '../../components/SendServiceRequestAttachment';
+import makeRequest                  from '../../util/requests';
+import serviceRequestModels         from '../../util/models';
+import Global                       from '../../util/globals';
+import Util                         from '../../util/util';
+import Config                       from '../../config';
+import transSendServiceRequest      from '../../translations/sendServiceRequest';
+import transError                   from '../../translations/errors';
+import checkboxIcon                 from '../../img/check.png';
+import sendEnabledIcon              from '../../img/send_enabled.png';
+import sendDisabledIcon             from '../../img/send_disabled.png';
+import markerIcon                   from '../../img/location_marker.png';
+import checkIcon                    from '../../img/check.png';
+import backIcon                     from '../../img/back.png';
+import styles                       from './styles';
 
-const BUTTON_ICON_HEIGHT     = 40;
-const BUTTON_ICON_WIDTH      = 40;
-const ZOOM                   = 6;
-const SEND_BUTTON_IMAGE_SIZE = 40;
+
+const ZOOM = 6;
 
 
 class SendServiceRequestView extends Component {
@@ -390,12 +387,17 @@ class SendServiceRequestView extends Component {
     LayoutAnimation.configureNext(LayoutAnimation.Presets.easeInEaseOut);
   }
 
+  onContentSizeChange = (event) => {
+    this.setState({
+      // Change the height of description view based on the length of the text
+      descriptionHeight: event.nativeEvent.contentSize.height
+    })
+  }
+
   onDescriptionChange = (event) => {
     if (event.nativeEvent.text.length < Config.OPEN311_DESCRIPTION_MAX_LENGTH) {
       this.setState({
-        descriptionText: event.nativeEvent.text,
-        // Change the height of description view based on the length of the text
-        descriptionHeight: event.nativeEvent.contentSize.height
+        descriptionText: event.nativeEvent.text
       });
     }
 
@@ -507,6 +509,7 @@ class SendServiceRequestView extends Component {
                 onTitleChange={this.onTitleChange}
                 descriptionHeight={this.state.descriptionHeight}
                 onDescriptionChange={this.onDescriptionChange}
+                onContentSizeChange={this.onContentSizeChange}
                 onPickerItemChange={this.onPickerItemChange}
                 checkboxImage={checkboxImage}
                 defaultTitle={this.state.titleText}
@@ -530,23 +533,5 @@ class SendServiceRequestView extends Component {
   }
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: Global.COLOR.LIGHT_GREY
-  },
-  innerContainer: {
-    flexDirection: 'column',
-    flex: 1,
-  },
-  scrollView: {
-    flex: 1,
-  },
-  checkboxImage: {
-    height: 32,
-    width: 32,
-  },
-
-});
 
 module.exports = SendServiceRequestView
